+++ docs/conf.py
@@ -3,8 +3,8 @@
 project = "seqtree"
 copyright = "2026, antigenomics"
 author = "antigenomics"
-release = "0.1.0"
-version = "0.1.0"
+release = "0.3.0"
+version = "0.3.0"
 
 extensions = [
     "sphinx.ext.autodoc",